Rename services to controllers
